requirements.txt
attrs
requests
click
flask
pydicom
pyyaml
redis
docker
pytest